[level 2] Title: 교점에 별 만들기, Time: 7.76 ms, Memory: 83.2 MB -BaekjoonHub

+++ 프로그래머스/2/87377. 교점에 별 만들기/교점에 별 만들기.java
@@ -14,10 +14,10 @@ class Solution {
                 if(xm!=0){
                     if(xj%xm==0&&yj%ym==0){
                         set.add((int)(xj/xm)+" "+(int)(yj/ym));
-                        xmin=(int)Math.min(xmin,(int)(xj/xm));
-                        xmax=(int)Math.max(xmax,(int)(xj/xm));
-                        ymin=(int)Math.min(ymin,(int)(yj/ym));
-                        ymax=(int)Math.max(ymax,(int)(yj/ym));
+                        xmin=Math.min(xmin,(int)(xj/xm));
+                        xmax=Math.max(xmax,(int)(xj/xm));
+                        ymin=Math.min(ymin,(int)(yj/ym));
+                        ymax=Math.max(ymax,(int)(yj/ym));
                     }
                 }  
             }
